feat(pa2): tensor/data parallel MPI comms + tiled matmul kernel
Part 2 (Tensor Parallel Communication, fully runnable on MS-MPI, CPU):
- comm.py: myAllreduce (reduce-to-root + broadcast) and myAlltoall (pairwise
  Sendrecv). Both verified 100/100 correct vs MPI built-ins on 8 ranks.
- data_parallel_preprocess.py: split_data across DP groups (model-parallel-major
  node layout). 4/4 tests pass.
- func_impl.py: get_info (mp/dp indexing + communicators + shard dims),
  naive_collect forward input/output (all-gather along feature axis) and
  backward output/x (split + reduce-scatter). get_info 2/2 (8 ranks),
  transformer forward 2/2 + backward 2/2 (4 ranks).
- mpi-test.py: fix Windows int32/int64 dtype mismatch in receive buffers.

Part 1 (MatMul+ReLU+Add kernel):
- matmul_triton.ipynb: complete Triton kernel (tile assignment, cooperative
  K-loop fetching, fp32 register accumulate, fused add+ReLU, masked epilogue
  store). Requires a CUDA/ROCm GPU (Colab T4 per writeup) so cannot run on this
  CPU box.
- matmul_tiled_cpu.py: CPU-runnable NumPy reference of the identical tiling
  algorithm; verified bit-close to reference across divisible and ragged dims.

+++ pa2/mpi-test.py
@@ -27,7 +27,7 @@ if __name__ == "__main__":
         """
 
         r = np.random.randint(0, 100, 100)
-        rr = np.empty(100, dtype=int)
+        rr = np.empty(100, dtype=r.dtype)  # match r's dtype (int32 on Windows)
 
         print("Rank " + str(rank) + ": " + str(r))
 
@@ -51,9 +51,10 @@ if __name__ == "__main__":
         for run in range(num_runs):
             # Create a random integer array (each process gets its own random array)
             r = np.random.randint(0, 100, 100)
-            # Arrays to store the reduction results
-            rr_allreduce   = np.empty(100, dtype=int)
-            rr_myallreduce = np.empty(100, dtype=int)
+            # Arrays to store the reduction results (match r's dtype for
+            # cross-platform correctness: randint yields int32 on Windows).
+            rr_allreduce   = np.empty(100, dtype=r.dtype)
+            rr_myallreduce = np.empty(100, dtype=r.dtype)
 
             # --- Built-in Allreduce Timing ---
             comm.Barrier()  # Synchronize all processes before timing
@@ -103,7 +104,7 @@ if __name__ == "__main__":
         """
 
         r = np.random.randint(0, 100, 2)
-        rr = np.empty(16, dtype=int)
+        rr = np.empty(16, dtype=r.dtype)
 
         print("Rank " + str(rank) + ": " + str(r))
 
@@ -119,7 +120,7 @@ if __name__ == "__main__":
         """
 
         r = np.random.randint(0, 100, 16)
-        rr = np.empty(2, dtype=int)
+        rr = np.empty(2, dtype=r.dtype)
 
         print("Rank " + str(rank) + ": " + str(r))
 
@@ -139,7 +140,7 @@ if __name__ == "__main__":
         """
 
         r = np.random.randint(0, 100, 10)
-        rr = np.empty(10, dtype=int)
+        rr = np.empty(10, dtype=r.dtype)
 
         print("Rank " + str(rank) + ": " + str(r))
